-Restructure the project -Move from returning bool to handle errors for repos -Add get all controller for images -Add delete controller for images Todo:    Alla:    -Update Image Controller    -Upload File

+++ src/Entity/Contract/Playlist.h
@@ -15,7 +15,7 @@ class Playlist : public IPlayable , public IObserver{
 private:
     int id;
     vector<unique_ptr<MediaDisplay>> display;
-    int current = 0;
+    unsigned long current = 0;
 
 public:
     Playlist(int id);
